Untangle web mercator, wgs84, vector tile implementations
