+++ benches/benchmarks/mod.rs
@@ -2,6 +2,6 @@ pub mod fenwick;
 pub mod great_circle;
 pub mod k_way_merge_iterator;
 pub mod loser_tree;
+pub mod mercator;
 pub mod polyline;
 pub mod radix_sort;
-pub mod wgs84;
